src/test/java/org/alan/algorithm/practice/baekjoon/clazz/five/BJ2473Test.java
package org.alan.algorithm.practice.baekjoon.clazz.five;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ2473Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        5
                        -2 6 -97 -6 98""", "-97 -2 98"),
                Arguments.of("""
                        7
                        -2 -3 -24 -6 98 100 61""", "-6 -3 -2")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ2473().solve(input));
    }
}
